better credential parsing from xml file..
Now, users don't have to define their polly project in the xml file..
they have an option "-N" in the CLI to specify polly project.
If this project is on polly, EPI will upload to that project else it will first create the project and then upload to it.
Also, better error propagation is there for credential parsing now..