Use boost library for decompression function
Decompression of zlib strings was being done using a c-style method
that was error prone (although worked well for the tested dataset).
Since we are trying to move away from raw buffer manipulation
techniques, we should start using more abstracted methods when
dealing with binary information. The Boost Iostreams library
provides a filter that can be used to deflate compressed data, and
will now be used for the decompression function in El-MAVEN.